Fix OpenCode loader crash by moving plugin test files
