+++ .opencode/lib/compat.js
