+++ package.json
@@ -12,7 +12,7 @@
     "README.md",
     "LICENSE",
     ".opencode/plugin/codemem.js",
-    ".opencode/plugin/compat.js"
+    ".opencode/lib/compat.js"
   ],
   "dependencies": {
     "@opencode-ai/plugin": "1.1.61"
